Updated LDBQueries to support Server port-binding
LDBQueries only had local data. Using the Server Class, we can integrate the server data that is incoming from the localhost webpage on port 8000. Doing this allows for an extra layer of protection, as the LDB Framework commands are lower level commands compared to the higher levels like server commands and client commands.